Cleaning up auth logic
