+++ src/test/java/net/getsentry/gocd/webhooknotifier/utils/HttpTest.java
@@ -22,7 +22,6 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.URL;
-import java.security.NoSuchAlgorithmException;
 
 public class HttpTest {
     @Test
@@ -35,38 +34,6 @@ public class HttpTest {
 
         verify(mockClient).execute(any(HttpPost.class));
         assertThat(response, is(mockResponse));
-    }
-
-    @Test
-    public void testGetAuthTokenException() throws IOException {
-        HttpClient mockClient = mock(HttpClient.class);
-        when(mockClient.execute(HttpGet.class.cast(any()))).thenThrow(new IOException());
-
-        String authToken = Http.getAuthToken(new URLWithAuth("https://example.com", "fakeAudience"), mockClient);
-
-        assertThat(authToken, is(nullValue()));
-    }
-
-    @Test
-    public void testGetAuthToken() throws IOException {
-        HttpClient mockClient = mock(HttpClient.class);
-        HttpResponse mockResponse = mock(HttpResponse.class);
-        HttpEntity mockHttpEntity = mock(HttpEntity.class);
-
-        when(mockResponse.getEntity()).thenReturn(mockHttpEntity);
-        when(mockHttpEntity.getContent()).thenReturn(new ByteArrayInputStream("fakeToken".getBytes()));
-        when(mockClient.execute(any())).thenReturn(mockResponse);
-
-        String authToken = Http.getAuthToken(new URLWithAuth("https://example.com", "fakeAudience"), mockClient);
-
-        verify(mockClient).execute(argThat((HttpGet request) -> {
-            try {
-                return request.getURI().toString().equals(Http.GCP_AUTH_METADATA_URL + "fakeAudience");
-            } catch (Exception e) {
-                return false;
-            }
-        }));
-        assertThat(authToken, is("fakeToken"));
     }
 
     @Test
@@ -193,11 +160,5 @@ public class HttpTest {
                 return false;
             }
         }));
-    }
-
-    @Test
-    public void testCreateSignature() throws NoSuchAlgorithmException {
-        String signature = Http.createSignature("test", "webhooksecret");
-        assertThat(signature, is("d01e0bee5d3b7783715126071e1c94df60bee516b4d6b1cb92896265cbbfe894"));
     }
 }
